requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.1
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.5
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.1.3
orjson==3.10.11
pandas==2.2.3
passlib==1.7.4
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg2==2.9.10
pydantic==2.9.2
pydantic-extra-types==2.10.0
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
rich==13.9.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.41.2
typer==0.13.0
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
uvicorn==0.32.0
watchfiles==0.24.0
websockets==14.1
